move FieldAccess stuff into compound.*
