+++ src/model/compound.h
@@ -23,6 +23,7 @@
 #include "expressions.h"
 #include "scope.h"
 #include "types.h"
+#include "semantics.h"
 
 #include <unordered_map>
 
@@ -47,6 +48,31 @@ public:
 
 private:
 	 EntryMap entries_;
+};
+
+/////////////////////////////////////////////////////////////////////////////////////////////////
+/***********************************************************************************************/
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+class FieldAccess
+: public Expression
+, public NoScopeOwner
+, public LanguageElement<FieldAccess>
+{
+public:
+	FieldAccess(Expression *subject, const string &field_name);
+	const Expression &subject() const;
+	const string &field_name() const;
+
+	DEFINE_ATTACH_SEMANTICS_WITH_MEMBERS(*subject_)
+
+	virtual const Type &type() const override;
+
+	string to_string(const string &pfx) const override;
+
+private:
+	SafeExprOwner<CompoundType> subject_;
+	const string field_name_;
 };
