+++ src/parser/effect_axiom.cpp
@@ -36,6 +36,7 @@
 #include <model/effect_axiom.h>
 #include <model/fluent.h>
 #include <model/list.h>
+#include <model/compound.h>
 
 #include "reference.h"
 #include "mixed_member_access.h"
