+++ src/semantics/readylog/compound.h
@@ -25,6 +25,25 @@
 namespace gologpp {
 
 
+std::pair<const Reference<Fluent> *, EC_word>
+traverse_mixed_field_access(const FieldAccess *fa, const ListAccess *la);
+
+
+template<>
+class Semantics<FieldAccess>
+: public Semantics<Expression>
+, public GeneralSemantics<FieldAccess>
+{
+public:
+	using GeneralSemantics<FieldAccess>::GeneralSemantics;
+
+	virtual EC_word plterm() override;
+	EC_word field_assign(const Expression &value);
+	EC_atom pl_field_name();
+};
+
+
+
 }
 
 #endif // READYLOG_COMPOUND_EXPRESSION_H_
